changed u16 random numbers to u32, seems collisions occur in tests (#1003)

+++ db/src/test/builders/artist_builder.rs
@@ -14,7 +14,7 @@ pub struct ArtistBuilder<'a> {
 
 impl<'a> ArtistBuilder<'a> {
     pub fn new(connection: &'a PgConnection) -> Self {
-        let x: u16 = random();
+        let x: u32 = random();
         ArtistBuilder {
             name: format!("Artist {}", x).into(),
             bio: "Bigraphy".into(),
